Added a new test for usage examples in readme. fixed error when invalid password is used

+++ BlockAppsSDK/Blocks/Block.cs
@@ -29,8 +29,8 @@ namespace BlockAppsSDK.Blocks
     public class BlockData
     {
         public int ExtraData { get; set; }
-        public int GasUsed { get; set; }
-        public int GasLimit { get; set; }
+        public double GasUsed { get; set; }
+        public double GasLimit { get; set; }
         public string Kind { get; set; }
         public string UnclesHash { get; set; }
         public string MixHash { get; set; }
@@ -40,7 +40,7 @@ namespace BlockAppsSDK.Blocks
         public string Timestamp { get; set; }
         public string Coinbase { get; set; }
         public string ParentHash { get; set; }
-        public int Nonce { get; set; }
+        public double Nonce { get; set; }
         public string StateRoot { get; set; }
         public string TransactionsRoot { get; set; }
     }
@@ -48,17 +48,17 @@ namespace BlockAppsSDK.Blocks
     public class Transaction
     {
         public string Hash { get; set; }
-        public int GasLimit { get; set; }
+        public double GasLimit { get; set; }
         public string Data { get; set; }
-        public int GasPrice { get; set; }
+        public double GasPrice { get; set; }
         public string CodeOrData { get; set; }
         public string To { get; set; }
-        public int Value { get; set; }
+        public double Value { get; set; }
         public string From { get; set; }
         public string r { get; set; }
         public string s { get; set; }
         public string v { get; set; }
-        public int Nonce { get; set; }
+        public double Nonce { get; set; }
     }
